requirements.txt

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0


langchain==1.0.3
langchain-openai==0.2.1
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langchain-core==0.2.38

# Pydantic (auto-resolve)
pydantic
pydantic-settings

# Vector Store
chromadb==0.4.22

# OpenAI
openai==1.10.0
tiktoken==0.5.2

# YouTube
youtube-transcript-api==0.6.1
pytube==15.0.0
yt-dlp==2023.12.30

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Web
tavily-python==0.5.0